# api/openapi.yaml
openapi: 3.1.0
info:
  title: Meraki Core Local API
  version: 0.2.0
  description: >-
    The authenticated local surface for immutable evidence, governed profile
    candidates, scoped runs, evaluations, trace inspection, and rollback.
security:
  - bearerAuth: []
paths:
  /dashboard:
    get:
      security: []
      operationId: getDashboard
      responses:
        "200":
          description: Local Meraki Studio HTML shell; API data remains bearer authenticated
          content:
            text/html:
              schema:
                type: string
  /health:
    get:
      security: []
      operationId: getHealth
      responses:
        "200":
          description: Process health and contract version
  /v1/corrections:
    post:
      operationId: recordCorrection
      responses:
        "201":
          description: Immutable correction evidence
        "403":
          description: Request identity does not match authenticated authority
        "422":
          description: Invalid correction
  /v1/activity:
    post:
      operationId: recordActivity
      responses:
        "201":
          description: Immutable explicit-user activity evidence
        "422":
          description: Invalid activity
  /v1/outcomes:
    post:
      operationId: recordOutcome
      responses:
        "201":
          description: Immutable objective-outcome evidence
        "422":
          description: Invalid outcome
  /v1/learning/candidates:
    post:
      operationId: createCandidate
      description: Creates an inactive candidate from authenticated evidence.
      responses:
        "201":
          description: Governed profile candidate
        "422":
          description: Evidence cannot support this candidate
  /v1/agent/run:
    post:
      operationId: runAgent
      responses:
        "200":
          description: Agent output, deterministic pack, and run trace
        "422":
          description: Invalid run request
  /v1/profile/atoms:
    get:
      operationId: listProfileAtoms
      responses:
        "200":
          description: Authenticated subject profile atoms
  /v1/profile/atoms/{id}/trace:
    get:
      operationId: getAtomTrace
      parameters:
        - $ref: "#/components/parameters/AtomId"
      responses:
        "200":
          description: Source-to-current-atom lineage
        "404":
          description: Atom trace not found
  /v1/profile/atoms/{id}/commands:
    post:
      operationId: commandProfileAtom
      parameters:
        - $ref: "#/components/parameters/AtomId"
      responses:
        "200":
          description: Versioned confirm, edit, rescope, limit, weaken, split, supersede, or revoke result
        "404":
          description: Atom not found
        "409":
          description: Version or lifecycle conflict
        "422":
          description: Invalid command
  /v1/learning/trace/{eventId}:
    get:
      operationId: getLearningTrace
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Source, event, observation, signal, hypothesis, and current atom
        "404":
          description: Learning trace not found
  /v1/update-proposals:
    get:
      operationId: listUpdateProposals
      responses:
        "200":
          description: Authenticated subject update proposals
    post:
      operationId: createUpdateProposal
      responses:
        "201":
          description: Pending evidence-backed update proposal
        "409":
          description: Target is not eligible for an update
        "422":
          description: Invalid proposal
  /v1/update-proposals/{id}/commands:
    post:
      operationId: commandUpdateProposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Applied, rejected, or rolled-back proposal
        "404":
          description: Proposal not found
        "409":
          description: Proposal lifecycle conflict
  /v1/runs:
    get:
      operationId: listRuns
      responses:
        "200":
          description: Authenticated subject run records
  /v1/runs/{runId}:
    get:
      operationId: getRun
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Run and complete trace
        "404":
          description: Run not found
  /v1/evaluations:
    get:
      operationId: listEvaluations
      responses:
        "200":
          description: Authenticated subject evaluations
    post:
      operationId: recordEvaluation
      responses:
        "201":
          description: Evaluator verdict and optional attribution
        "404":
          description: Run not found
        "422":
          description: Invalid evaluation
  /v1/evaluations/causal:
    post:
      operationId: runControlledComparison
      description: >-
        Deterministic showcase comparison. Provider-backed experiments use the
        asynchronous executor surface and are recorded separately.
      responses:
        "201":
          description: Baseline, raw-memory, Meraki, and ablation report
        "422":
          description: Invalid experiment
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AtomId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid